contrib: aws-vault proxy must listen on 169.254.170.2
Since this endpoint exposes AWS credentials directly to any callers, it
is of utmost importance that it not be called by unauthorized 3rd
parties. By listening only on 169.254.170.2, we ensure that it's only
possible for callers that have access to that IP address to retrieve
credentials. This avoids the potential mistake of e.g. hosting this with
host networking and allowing it publicly on all available interfaces.